Fix: フラッシュメッセージをコンポーネント化した
