+++ resources/scss/global/constants/_variable.scss
@@ -12,5 +12,7 @@ $light-black: #333333;
 $light-gray: #dadce0;
 $light-blue: rgb(96, 165, 250);
 $blue: rgb(37, 99, 235);
-$green: rgb(187, 247, 208);
-$red: rgb(243, 113, 113);
+$light-green: rgb(187, 247, 208);
+$green: rgb(74, 222, 128);
+$light-red: rgb(254, 202, 202);
+$red: rgb(248, 113, 113);
